Improved error messaging for clipboard access failure due to website not being accessed over secure connection.
We now hide the copy buttons and pop up an alert ahead of time, as well
a log a warning in the console.

We also added improved error handling in the button events, but we exit
out early so they shouldn't be needed. Left in in case we don't want to
early out some time in the future.